chore(Hotjar): move sdk loader to different file
