+++ src/integrations/Hotjar/browser.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable class-methods-use-this */
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
+import { loader } from './loader';
 
 class Hotjar {
   constructor(config, analytics, destinationInfo) {
@@ -22,20 +22,7 @@ class Hotjar {
     logger.debug('===In init Hotjar===');
 
     window.hotjarSiteId = this.siteId;
-    (function (h, o, t, j, a, r) {
-      h.hj =
-        h.hj ||
-        function () {
-          (h.hj.q = h.hj.q || []).push(arguments);
-        };
-      h._hjSettings = { hjid: h.hotjarSiteId, hjsv: 6 };
-      a = o.getElementsByTagName('head')[0];
-      r = o.createElement('script');
-      r.setAttribute('data-loader', LOAD_ORIGIN);
-      r.async = 1;
-      r.src = t + h._hjSettings.hjid + j + h._hjSettings.hjsv;
-      a.appendChild(r);
-    })(window, document, 'https://static.hotjar.com/c/hotjar-', '.js?sv=');
+    loader(window, document, 'https://static.hotjar.com/c/hotjar-', '.js?sv=');
     this._ready = true;
   }
